requirements.txt
torch==2.1.1
diffusers==0.28.0
transformers==4.41.2
accelerate==0.30.1
huggingface_hub==0.23.4
Pillow
requests
Flask
opencv-python
imageio
imageio-ffmpeg
numpy<2.0
basicsr
realesrgan
ffmpeg-python
facenet-pytorch
peft